+++ CHANGELOG.md
@@ -1,9 +1,16 @@
-This document lists the changes to `opencap-core`. When possible, we provide the date, GitHub issues, or pull requests that are related to the items below. If there is no issue or pull request related to the change, then we may provide the commit.
+This document lists the changes to `opencap-core` for each version. When possible, we provide the GitHub issues or pull requests that are related to the items below. If there is no issue or pull request related to the change, then we may provide the commit.
 
 This is not a comprehensive list of changes but rather a hand-curated collection of the more notable ones. For a comprehensive history, see the [OpenCap Core GitHub repo](https://github.com/stanfordnmbl/opencap-core).
 
-Changes
-=======
+v1.1
+=====
+- Improved synchronization with an arm raise (hand punch). (#182)
+- Added option for downloading pose pickle files. (#248)
+- Moved synchronization specific files to new module utilsSync.py. (PR #259)
+- Added main regression tests and sync unit tests. (PR #259)
+
+Previous Changes
+================
 - 07/03/2024: Speed up IK by removing patella constraints from model ([pull request](https://github.com/stanfordnmbl/opencap-core/pull/174))
 - 06/25/2024: Add support for "any pose" scaling ([pull request](https://github.com/stanfordnmbl/opencap-core/pull/168))
 - 05/10/2024: Add reprojection error minimization to improve camera synchronization  ([pull request](https://github.com/stanfordnmbl/opencap-core/pull/159))